refactor(cli): daemon サブコマンドを daemon/cli.py に切り出す
cli.py をコマンド登録のみに絞り、daemon サブグループの
ヘルパー関数とサブコマンド定義を meeting_minutes.daemon.cli へ移動した。
AGENTS.md に「cli.py はコマンド登録のみ」の原則を追記した。
